# Conflicts:
#	Modules/Physics/CMakeLists.txt
#	Sources/Internal/CMakeLists.txt


Former-commit-id: 90fd8418a6ffd867c3fff5f016d3f166d478696d

+++ Sources/Internal/Autotesting/AutotestingSystem.h
@@ -157,7 +157,7 @@ private:
 
 public:
     static const String RecordScriptFileName;
-    float32 startTime = 0.f;
+    int64 startTime = 0;
 
     bool isInit;
     bool isRunning;
